+++ user/src/main/java/com/owlexpress/user/common/constant/Role.java
@@ -1,4 +1,4 @@
-package com.owlexpress.user.domain.constant;
+package com.owlexpress.user.common.constant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
