+++ user/src/main/java/com/owlexpress/user/common/dto/PassportDto.java
@@ -1,6 +1,6 @@
 package com.owlexpress.user.common.dto;
 
-import com.owlexpress.user.domain.constant.Role;
+import com.owlexpress.user.common.constant.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
