+++ user/src/main/java/com/owlexpress/user/common/exception/NewPasswordNotMatchesException.java
@@ -1,6 +1,4 @@
-package com.owlexpress.user.application.exception;
-
-import com.owlexpress.user.common.exception.AbstractUserUpdateException;
+package com.owlexpress.user.common.exception;
 
 public class NewPasswordNotMatchesException extends AbstractUserUpdateException {
     public NewPasswordNotMatchesException(String message) {
