+++ user/src/main/java/com/owlexpress/user/common/exception/PasswordIncorrectException.java
@@ -1,4 +1,4 @@
-package com.owlexpress.user.application.exception;
+package com.owlexpress.user.common.exception;
 
 public class PasswordIncorrectException extends RuntimeException{
     public PasswordIncorrectException(String message) {super(message);};
