+++ user/src/main/java/com/owlexpress/user/presentation/dto/request/ApprovalUserRequestDto.java
@@ -1,6 +1,6 @@
 package com.owlexpress.user.presentation.dto.request;
 
-import com.owlexpress.user.domain.constant.ApprovalStatus;
+import com.owlexpress.user.common.constant.ApprovalStatus;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
