+++ user/src/main/java/com/owlexpress/user/presentation/dto/response/GetUserInfoResponseDto.java
@@ -1,7 +1,7 @@
 package com.owlexpress.user.presentation.dto.response;
 
-import com.owlexpress.user.domain.constant.PlatformType;
-import com.owlexpress.user.domain.constant.Role;
+import com.owlexpress.user.common.constant.PlatformType;
+import com.owlexpress.user.common.constant.Role;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
